+++ cmp/tab/TabContainerModel.js
@@ -51,8 +51,7 @@ export class TabContainerModel {
      *      be route-enabled, with the route for each tab being "[route]/[tab.id]".
      * @param {string} [c.switcherPosition] - Position of the switcher docked within this component (or 'none').
      *      Valid values are 'top', 'bottom', 'left', 'right', 'none'.
-     * @param {boolean} [c.track] - Enables tracking of tab views in admin logger under category
-     *      "Tabs". If not set, will default to false.
+     * @param {boolean} [c.track] - True to enable activity tracking of tab views. The default value is set to 'false.'
      * @param {RenderMode} [c.renderMode] - strategy for rendering child tabs. Can be set
      *      per-tab via `TabModel.renderMode`. See enum for description of supported modes.
      * @param {RefreshMode} [c.refreshMode] - strategy for refreshing child tabs. Can be set
@@ -99,10 +98,10 @@ export class TabContainerModel {
             this.addReaction({
                 track: () => this.activeTab,
                 run: (activeTab) => {
-                    const route = activeTab.containerModel.route;
+                    const {route} = this;
                     XH.track({
-                        msg: `Viewed ${activeTab.title}`,
-                        category: 'Tab',
+                        category: 'Navigation',
+                        message: `Viewed ${activeTab.title} tab`,
                         // If using routing, data field specifies route for non-top-level tabs.
                         data: route && route !== 'default' ? {route: route} : null
                     });